Move commands from pkg to internal
