+++ internal/cli/cmd/canary/install.go
