+++ internal/cli/cmd/canary/uninstall.go
