+++ internal/cli/cmd/demoapp/uninstall.go
