+++ internal/cli/cmd/routing/cb/cmd.go
@@ -12,26 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package routing
+package cb
 
 import (
 	"github.com/spf13/cobra"
 
+	"github.com/banzaicloud/backyards-cli/internal/cli/cmd/graph"
 	"github.com/banzaicloud/backyards-cli/pkg/cli"
-	"github.com/banzaicloud/backyards-cli/pkg/cli/cmd/routing/cb"
-	"github.com/banzaicloud/backyards-cli/pkg/cli/cmd/routing/ts"
 )
 
 func NewRootCmd(cli cli.CLI) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:     "routing",
-		Aliases: []string{"r"},
-		Short:   "Manage service routing configurations",
+		Use:     "circuit-breaker",
+		Aliases: []string{"cb"},
+		Short:   "Manage circuit-breaker configurations",
 	}
 
 	cmd.AddCommand(
-		ts.NewRootCmd(cli),
-		cb.NewRootCmd(cli),
+		newGetCommand(cli),
+		newSetCommand(cli),
+		newDeleteCommand(cli),
+		graph.NewGraphCmd(cli, "cb.json"),
 	)
 
 	return cmd
