+++ internal/cli/cmd/uninstall.go
@@ -22,11 +22,11 @@ import (
 	"github.com/spf13/cobra"
 	"k8s.io/apimachinery/pkg/util/wait"
 
+	"github.com/banzaicloud/backyards-cli/internal/cli/cmd/canary"
+	"github.com/banzaicloud/backyards-cli/internal/cli/cmd/certmanager"
+	"github.com/banzaicloud/backyards-cli/internal/cli/cmd/demoapp"
+	"github.com/banzaicloud/backyards-cli/internal/cli/cmd/istio"
 	"github.com/banzaicloud/backyards-cli/pkg/cli"
-	"github.com/banzaicloud/backyards-cli/pkg/cli/cmd/canary"
-	"github.com/banzaicloud/backyards-cli/pkg/cli/cmd/certmanager"
-	"github.com/banzaicloud/backyards-cli/pkg/cli/cmd/demoapp"
-	"github.com/banzaicloud/backyards-cli/pkg/cli/cmd/istio"
 	"github.com/banzaicloud/backyards-cli/pkg/helm"
 	"github.com/banzaicloud/backyards-cli/pkg/k8s"
 )
@@ -45,7 +45,7 @@ type UninstallOptions struct {
 	uninstallEverything  bool
 }
 
-func newUninstallCommand(cli cli.CLI) *cobra.Command {
+func NewUninstallCommand(cli cli.CLI) *cobra.Command {
 	c := &uninstallCommand{}
 	options := &UninstallOptions{}
